Changes & improvements from first round of PR review.
