+++ api/src/integration/java/org/pmiops/workbench/IntegrationTestConfig.java
@@ -10,10 +10,6 @@ import org.pmiops.workbench.firecloud.FireCloudConfig;
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Import;
-import org.springframework.retry.backoff.BackOffPolicy;
-import org.springframework.retry.backoff.ExponentialRandomBackOffPolicy;
-import org.springframework.retry.backoff.Sleeper;
-import org.springframework.retry.backoff.ThreadWaitSleeper;
 
 @TestConfiguration
 @Import({RetryConfig.class, CommonConfig.class, StoredCredentialsConfig.class})
@@ -26,7 +22,7 @@ public class IntegrationTestConfig {
   }
 
   /**
-   * Returns the Apache HTTP transport. Compare to CommonConfig which returns the App Engine HTTP
+   * Returns the Apache HTTP transport. Compare to AppEngineConfig which returns the App Engine HTTP
    * transport.
    *
    * @return
@@ -34,20 +30,5 @@ public class IntegrationTestConfig {
   @Bean
   HttpTransport httpTransport() {
     return new ApacheHttpTransport();
-  }
-
-  @Bean
-  public Sleeper sleeper() {
-    return new ThreadWaitSleeper();
-  }
-
-  @Bean
-  public BackOffPolicy backOffPolicy(Sleeper sleeper) {
-    // Defaults to 100ms initial interval, doubling each time, with some random multiplier.
-    ExponentialRandomBackOffPolicy policy = new ExponentialRandomBackOffPolicy();
-    // Set max interval of 20 seconds.
-    policy.setMaxInterval(20000);
-    policy.setSleeper(sleeper);
-    return policy;
   }
 }
