+++ api/src/main/java/org/pmiops/workbench/firecloud/FireCloudConfig.java
@@ -1,5 +1,6 @@
 package org.pmiops.workbench.firecloud;
 
+import com.google.auth.oauth2.GoogleCredentials;
 import com.google.common.collect.ImmutableList;
 import java.io.IOException;
 import java.util.List;
@@ -57,7 +58,9 @@ public class FireCloudConfig {
   public ApiClient allOfUsApiClient(WorkbenchConfig workbenchConfig) {
     ApiClient apiClient = buildApiClient(workbenchConfig);
     try {
-      apiClient.setAccessToken(ServiceAccounts.workbenchAccessToken(BILLING_SCOPES));
+      apiClient.setAccessToken(
+          ServiceAccounts.getScopedAccessToken(
+              GoogleCredentials.getApplicationDefault(), BILLING_SCOPES));
     } catch (IOException e) {
       throw new ServerErrorException(e);
     }
