+++ api/src/main/java/org/pmiops/workbench/firecloud/FireCloudService.java
@@ -1,5 +1,6 @@
 package org.pmiops.workbench.firecloud;
 
+import com.google.common.annotations.VisibleForTesting;
 import java.io.IOException;
 import java.util.List;
 import java.util.Optional;
@@ -25,7 +26,8 @@ public interface FireCloudService {
   String WORKSPACE_DELIMITER = "__";
 
   /** Returns the base path for the FireCloud API. Exposed for integration testing purposes only. */
-  String getBasePath();
+  @VisibleForTesting
+  String getApiBasePath();
 
   /** @return true if firecloud is okay, false if firecloud is down. */
   boolean getFirecloudStatus();
