+++ api/src/main/java/org/pmiops/workbench/firecloud/FirecloudRetryHandler.java
@@ -3,9 +3,9 @@ package org.pmiops.workbench.firecloud;
 import java.net.SocketTimeoutException;
 import java.util.logging.Logger;
 import javax.servlet.http.HttpServletResponse;
-import org.pmiops.workbench.config.RetryConfig;
 import org.pmiops.workbench.exceptions.ExceptionUtils;
 import org.pmiops.workbench.exceptions.WorkbenchException;
+import org.pmiops.workbench.utils.ResponseCodeRetryPolicy;
 import org.pmiops.workbench.utils.RetryHandler;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.retry.backoff.BackOffPolicy;
@@ -16,7 +16,7 @@ public class FirecloudRetryHandler extends RetryHandler<ApiException> {
 
   private static final Logger logger = Logger.getLogger(FirecloudRetryHandler.class.getName());
 
-  private static class FirecloudRetryPolicy extends RetryConfig.ResponseCodeRetryPolicy {
+  private static class FirecloudRetryPolicy extends ResponseCodeRetryPolicy {
 
     public FirecloudRetryPolicy() {
       super("Firecloud API");
