+++ api/src/main/java/org/pmiops/workbench/notebooks/NotebooksConfig.java
@@ -1,5 +1,6 @@
 package org.pmiops.workbench.notebooks;
 
+import com.google.auth.oauth2.GoogleCredentials;
 import com.google.common.collect.ImmutableList;
 import java.io.IOException;
 import java.util.List;
@@ -42,7 +43,9 @@ public class NotebooksConfig {
   public ApiClient workbenchServiceAccountClient(WorkbenchConfig workbenchConfig) {
     ApiClient apiClient = buildApiClient(workbenchConfig);
     try {
-      apiClient.setAccessToken(ServiceAccounts.workbenchAccessToken(NOTEBOOK_SCOPES));
+      apiClient.setAccessToken(
+          ServiceAccounts.getScopedAccessToken(
+              GoogleCredentials.getApplicationDefault(), NOTEBOOK_SCOPES));
     } catch (IOException e) {
       throw new ServerErrorException(e);
     }
