+++ api/src/main/java/org/pmiops/workbench/notebooks/NotebooksRetryHandler.java
@@ -3,9 +3,9 @@ package org.pmiops.workbench.notebooks;
 import java.net.SocketTimeoutException;
 import java.util.logging.Logger;
 import javax.servlet.http.HttpServletResponse;
-import org.pmiops.workbench.config.RetryConfig;
 import org.pmiops.workbench.exceptions.ExceptionUtils;
 import org.pmiops.workbench.exceptions.WorkbenchException;
+import org.pmiops.workbench.utils.ResponseCodeRetryPolicy;
 import org.pmiops.workbench.utils.RetryHandler;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.retry.backoff.BackOffPolicy;
@@ -16,7 +16,7 @@ public class NotebooksRetryHandler extends RetryHandler<ApiException> {
 
   private static final Logger logger = Logger.getLogger(NotebooksRetryHandler.class.getName());
 
-  private static class NotebookRetryPolicy extends RetryConfig.ResponseCodeRetryPolicy {
+  private static class NotebookRetryPolicy extends ResponseCodeRetryPolicy {
 
     public NotebookRetryPolicy() {
       super("Notebooks API");
